+++ pydash/__init__.py
@@ -326,6 +326,7 @@ from .predicates import (
     is_json,
     is_list,
     is_match,
+    is_match_with,
     is_monotone,
     is_nan,
     is_native,