+++ tests/test_predicates.py
@@ -373,10 +373,6 @@ def test_is_list(case, expected):
     (({'name': 'fred', 'age': 40}, {'age': 40, 'active': True}), False),
     (([1, 2, 3], [1, 2]), True),
     (([1, 2, 3], [1, 2, 3, 4]), False),
-    (([1, 2],
-      [2, 4],
-      lambda a, b: None if isinstance(a, list) else b == a + a),
-     True),
     (({}, {}), True),
     (({'a': 1}, {}), True),
     (([], []), True),
@@ -384,6 +380,24 @@ def test_is_list(case, expected):
 ])
 def test_is_match(case, expected):
     assert _.is_match(*case) == expected
+
+
+@parametrize('case,expected', [
+    (([1, 2],
+      [2, 4],
+      lambda a, b: None if isinstance(a, list) else b == a + a),
+     True),
+    (([1, 2],
+      [2, 4],
+      lambda a, b, key: a == b == key),
+     False),
+    (([0, 1],
+      [0, 1],
+      lambda a, b, key: a == b == key),
+     True),
+])
+def test_is_match_with(case, expected):
+    assert _.is_match_with(*case) == expected
 
 
 @parametrize('case,expected', [